[UPD] improve git clone performances
dear maintainers,

i reviewed the code of this script and i discovered that git clone works by cloning the current state of the repository and also all the history. i think that someone using searchsploit -u will not use the history, but only the current state.

so i propose you to use the option "--depth=1", which allows the clone of only the current state of the db.

if you think this is a bad idea for any reason, feel free to drop this request (if so, could you please let me know why? just because i am curious :D ).

thanks a lot for providing exploit-db, i really appreciate the effort and the project itself!

happy hacking!

Lo